+++ src/main/java/mil/nga/wkb/util/GeometryUtils.java
@@ -1,0 +1,110 @@
+package mil.nga.wkb.util;
+
+import java.util.List;
+
+import mil.nga.wkb.geom.Geometry;
+import mil.nga.wkb.geom.GeometryCollection;
+import mil.nga.wkb.geom.GeometryType;
+import mil.nga.wkb.geom.Point;
+
+/**
+ * Utilities for Geometry objects
+ * 
+ * @author osbornb
+ * @since 1.0.3
+ */
+public class GeometryUtils {
+
+	/**
+	 * Get the dimension of the Geometry, 0 for points, 1 for curves, 2 for
+	 * surfaces. If a collection, the largest dimension is returned.
+	 * 
+	 * @param geometry
+	 *            geometry object
+	 * @return dimension (0, 1, or 2)
+	 */
+	public static int getDimension(Geometry geometry) {
+
+		int dimension = -1;
+
+		GeometryType geometryType = geometry.getGeometryType();
+		switch (geometryType) {
+		case POINT:
+		case MULTIPOINT:
+			dimension = 0;
+			break;
+		case LINESTRING:
+		case MULTILINESTRING:
+		case CIRCULARSTRING:
+		case COMPOUNDCURVE:
+			dimension = 1;
+			break;
+		case POLYGON:
+		case CURVEPOLYGON:
+		case MULTIPOLYGON:
+		case POLYHEDRALSURFACE:
+		case TIN:
+		case TRIANGLE:
+			dimension = 2;
+			break;
+		case GEOMETRYCOLLECTION:
+			@SuppressWarnings("unchecked")
+			GeometryCollection<Geometry> geomCollection = (GeometryCollection<Geometry>) geometry;
+			List<Geometry> geometries = geomCollection.getGeometries();
+			for (Geometry subGeometry : geometries) {
+				dimension = Math.max(dimension, getDimension(subGeometry));
+			}
+			break;
+		default:
+			throw new WkbException("Unsupported Geometry Type: " + geometryType);
+		}
+
+		return dimension;
+	}
+
+	/**
+	 * Get the Pythagorean theorem distance between two points
+	 * 
+	 * @param point1
+	 *            point 1
+	 * @param point2
+	 *            point 2
+	 * @return distance
+	 */
+	public static double distance(Point point1, Point point2) {
+		double diffX = point1.getX() - point2.getX();
+		double diffY = point1.getY() - point2.getY();
+
+		double distance = Math.sqrt(diffX * diffX + diffY * diffY);
+
+		return distance;
+	}
+
+	/**
+	 * Get the centroid point of the Geometry
+	 * 
+	 * @param geometry
+	 *            geometry object
+	 * @return centroid point
+	 */
+	public static Point getCentroid(Geometry geometry) {
+		Point centroid = null;
+		int dimension = getDimension(geometry);
+		switch (dimension) {
+		case 0:
+			CentroidPoint point = new CentroidPoint(geometry);
+			centroid = point.getCentroid();
+			break;
+		case 1:
+			CentroidCurve curve = new CentroidCurve(geometry);
+			centroid = curve.getCentroid();
+			break;
+		case 2:
+			CentroidSurface surface = new CentroidSurface(geometry);
+			centroid = surface.getCentroid();
+			break;
+		}
+		return centroid;
+	}
+
+}