+++ src/test/java/mil/nga/wkb/test/GeometryUtilsTest.java
@@ -1,0 +1,273 @@
+package mil.nga.wkb.test;
+
+import java.io.IOException;
+
+import junit.framework.TestCase;
+import mil.nga.wkb.geom.Geometry;
+import mil.nga.wkb.geom.GeometryCollection;
+import mil.nga.wkb.geom.GeometryEnvelope;
+import mil.nga.wkb.geom.GeometryType;
+import mil.nga.wkb.geom.LineString;
+import mil.nga.wkb.geom.MultiLineString;
+import mil.nga.wkb.geom.MultiPoint;
+import mil.nga.wkb.geom.MultiPolygon;
+import mil.nga.wkb.geom.Point;
+import mil.nga.wkb.geom.Polygon;
+import mil.nga.wkb.util.GeometryEnvelopeBuilder;
+import mil.nga.wkb.util.GeometryUtils;
+
+import org.junit.Test;
+
+/**
+ * Test Well Known Binary Geometry Utilities
+ * 
+ * @author osbornb
+ */
+public class GeometryUtilsTest {
+
+	/**
+	 * Number of random geometries to create for each test
+	 */
+	private static final int GEOMETRIES_PER_TEST = 10;
+
+	/**
+	 * Constructor
+	 */
+	public GeometryUtilsTest() {
+
+	}
+
+	@Test
+	public void testPointCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a point
+			Point point = WKBTestUtils.createPoint(WKBTestUtils.coinFlip(),
+					WKBTestUtils.coinFlip());
+			TestCase.assertEquals(0, GeometryUtils.getDimension(point));
+			geometryCentroidTester(point);
+		}
+
+	}
+
+	@Test
+	public void testLineStringCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a line string
+			LineString lineString = WKBTestUtils.createLineString(
+					WKBTestUtils.coinFlip(), WKBTestUtils.coinFlip());
+			TestCase.assertEquals(1, GeometryUtils.getDimension(lineString));
+			geometryCentroidTester(lineString);
+		}
+
+	}
+
+	@Test
+	public void testPolygonCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a polygon
+			Polygon polygon = createPolygon();
+			TestCase.assertEquals(2, GeometryUtils.getDimension(polygon));
+			geometryCentroidTester(polygon);
+		}
+
+	}
+
+	@Test
+	public void testMultiPointCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a multi point
+			MultiPoint multiPoint = WKBTestUtils.createMultiPoint(
+					WKBTestUtils.coinFlip(), WKBTestUtils.coinFlip());
+			TestCase.assertEquals(0, GeometryUtils.getDimension(multiPoint));
+			geometryCentroidTester(multiPoint);
+		}
+
+	}
+
+	@Test
+	public void testMultiLineStringCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a multi line string
+			MultiLineString multiLineString = WKBTestUtils
+					.createMultiLineString(WKBTestUtils.coinFlip(),
+							WKBTestUtils.coinFlip());
+			TestCase.assertEquals(1,
+					GeometryUtils.getDimension(multiLineString));
+			geometryCentroidTester(multiLineString);
+		}
+
+	}
+
+	@Test
+	public void testMultiPolygonCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a multi polygon
+			MultiPolygon multiPolygon = createMultiPolygon();
+			TestCase.assertEquals(2, GeometryUtils.getDimension(multiPolygon));
+			geometryCentroidTester(multiPolygon);
+		}
+
+	}
+
+	@Test
+	public void testGeometryCollectionCentroid() throws IOException {
+
+		for (int i = 0; i < GEOMETRIES_PER_TEST; i++) {
+			// Create and test a geometry collection
+			GeometryCollection<Geometry> geometryCollection = createGeometryCollection(
+					WKBTestUtils.coinFlip(), WKBTestUtils.coinFlip());
+			geometryCentroidTester(geometryCollection);
+		}
+
+	}
+
+	@Test
+	public void testPolygonCentroidWithAndWithoutHole() throws IOException {
+
+		Polygon polygon = new Polygon();
+		LineString lineString = new LineString();
+		lineString.addPoint(new Point(-90, 45));
+		lineString.addPoint(new Point(-90, -45));
+		lineString.addPoint(new Point(90, -45));
+		lineString.addPoint(new Point(90, 45));
+		polygon.addRing(lineString);
+
+		TestCase.assertEquals(2, GeometryUtils.getDimension(polygon));
+		Point centroid = geometryCentroidTester(polygon);
+
+		TestCase.assertEquals(0.0, centroid.getX());
+		TestCase.assertEquals(0.0, centroid.getY());
+
+		LineString holeLineString = new LineString();
+		holeLineString.addPoint(new Point(0, 45));
+		holeLineString.addPoint(new Point(0, 0));
+		holeLineString.addPoint(new Point(90, 0));
+		holeLineString.addPoint(new Point(90, 45));
+		polygon.addRing(holeLineString);
+
+		TestCase.assertEquals(2, GeometryUtils.getDimension(polygon));
+		centroid = geometryCentroidTester(polygon);
+
+		TestCase.assertEquals(-15.0, centroid.getX());
+		TestCase.assertEquals(-7.5, centroid.getY());
+	}
+
+	/**
+	 * Test the geometry centroid
+	 * 
+	 * @param geometry
+	 * @throws IOException
+	 */
+	private Point geometryCentroidTester(Geometry geometry) throws IOException {
+
+		Point point = GeometryUtils.getCentroid(geometry);
+
+		GeometryEnvelope envelope = GeometryEnvelopeBuilder
+				.buildEnvelope(geometry);
+
+		if (geometry.getGeometryType() == GeometryType.POINT) {
+			TestCase.assertEquals(envelope.getMinX(), point.getX());
+			TestCase.assertEquals(envelope.getMaxX(), point.getX());
+			TestCase.assertEquals(envelope.getMinY(), point.getY());
+			TestCase.assertEquals(envelope.getMaxY(), point.getY());
+		}
+
+		TestCase.assertTrue(point.getX() >= envelope.getMinX());
+		TestCase.assertTrue(point.getX() <= envelope.getMaxX());
+		TestCase.assertTrue(point.getY() >= envelope.getMinY());
+		TestCase.assertTrue(point.getY() <= envelope.getMaxY());
+
+		return point;
+	}
+
+	private static Polygon createPolygon() {
+
+		Polygon polygon = new Polygon();
+		LineString lineString = new LineString();
+		lineString.addPoint(createPoint(-180.0, 45.0, 90.0, 45.0));
+		lineString.addPoint(createPoint(-180.0, -90.0, 90.0, 45.0));
+		lineString.addPoint(createPoint(90.0, -90.0, 90.0, 45.0));
+		lineString.addPoint(createPoint(90.0, 45.0, 90.0, 45.0));
+		polygon.addRing(lineString);
+
+		LineString holeLineString = new LineString();
+		holeLineString.addPoint(createPoint(-90.0, 0.0, 90.0, 45.0));
+		holeLineString.addPoint(createPoint(-90.0, -45.0, 90.0, 45.0));
+		holeLineString.addPoint(createPoint(0.0, -45.0, 90.0, 45.0));
+		holeLineString.addPoint(createPoint(0.0, 0.0, 90.0, 45.0));
+		polygon.addRing(holeLineString);
+
+		return polygon;
+	}
+
+	private static Point createPoint(double minX, double minY, double xRange,
+			double yRange) {
+
+		double x = minX + (Math.random() * xRange);
+		double y = minY + (Math.random() * yRange);
+
+		Point point = new Point(x, y);
+
+		return point;
+	}
+
+	private static MultiPolygon createMultiPolygon() {
+
+		MultiPolygon multiPolygon = new MultiPolygon();
+
+		int num = 1 + ((int) (Math.random() * 5));
+
+		for (int i = 0; i < num; i++) {
+			multiPolygon.addPolygon(createPolygon());
+		}
+
+		return multiPolygon;
+	}
+
+	private static GeometryCollection<Geometry> createGeometryCollection(
+			boolean hasZ, boolean hasM) {
+
+		GeometryCollection<Geometry> geometryCollection = new GeometryCollection<Geometry>(
+				hasZ, hasM);
+
+		int num = 1 + ((int) (Math.random() * 5));
+
+		for (int i = 0; i < num; i++) {
+
+			Geometry geometry = null;
+			int randomGeometry = (int) (Math.random() * 6);
+
+			switch (randomGeometry) {
+			case 0:
+				geometry = WKBTestUtils.createPoint(hasZ, hasM);
+				break;
+			case 1:
+				geometry = WKBTestUtils.createLineString(hasZ, hasM);
+				break;
+			case 2:
+				geometry = createPolygon();
+				break;
+			case 3:
+				geometry = WKBTestUtils.createMultiPoint(hasZ, hasM);
+				break;
+			case 4:
+				geometry = WKBTestUtils.createMultiLineString(hasZ, hasM);
+				break;
+			case 5:
+				geometry = createMultiPolygon();
+				break;
+			}
+
+			geometryCollection.addGeometry(geometry);
+		}
+
+		return geometryCollection;
+	}
+
+}