+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 1.0.3 (TBD)
 
 * Shortcut default constructors for Geometry objects without z or m values
+* Geometry utilities including dimension and centroid
 
 ## [1.0.2](https://github.com/ngageoint/geopackage-wkb-java/releases/tag/1.0.2) (04-18-2016)